+++ src/main/resources/js/scrum-poker.js
@@ -55,9 +55,16 @@
                 $('#scrum-poker-finished').empty();
                 cancelRedirectToIssue();
             }
+            customStylingForCards();
         }).fail(function() {
             scheduleRedirectToIssue(issueKey);
         });
+    }
+
+    /* special cards like the question mark and the coffee card are rendered with an icon */
+    function customStylingForCards() {
+        $('.card-value-coffee').addClass('fas fa-mug-hot').empty();
+        $('.card-value-question').addClass('fas fa-question').empty();
     }
 
     /* Schedule a redirect to the issue page after 60 seconds */