+++ docs/changelog.md
@@ -13,6 +13,7 @@ Click on the version in the following list to see all changes since the previous
 
 * feat: signal that a break is needed only when deck is revealed
 * docs: describe possible usage of test groups
+* feat: use Font Awesome to display icons for coffee card and question mark
 
 ### [3.11.1] - 2019-04-01